# specs/001-trading-wizard-ui/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Trading Wizard API
  version: 1.0.0
  description: |
    Trading Wizard 웹 인터페이스 API.
    - Daily Focus Wizard: 매수 추천 종목 조회
    - My Portfolio Wizard: 포트폴리오 관리
    - 공통: 사용자 데이터 저장/조회 (암호화된 blob)
servers:
  - url: http://localhost:8000
    description: Development server
...
paths:
...
  /api/stocks/search:
    get:
      tags:
        - Stock
      summary: 종목 검색
      description: 종목코드 또는 종목명으로 검색합니다.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: 검색어 (종목코드 또는 종목명)
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
          description: 최대 결과 수
      responses:
        '200':
          description: 검색 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockSearchResponse'
...
  /api/stocks/{symbol}/price:
    get:
      tags:
        - Stock
      summary: 종목 현재가 조회
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
          example: "005930.KS"
      responses:
        '200':
          description: 종목 현재가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockPriceResponse'
        '404':
          description: 종목을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 사용자 친화적 에러 메시지
...
    Stock:
      type: object
      required:
        - symbol
        - name
        - market
        - currentPrice
      properties:
        symbol:
          type: string
          example: "005930.KS"
        name:
          type: string
          example: "삼성전자"
        market:
          type: string
          enum: [KOSPI, KOSDAQ]
        currentPrice:
          type: number
          example: 72500
        previousClose:
          type: number
          example: 72000
        changePercent:
          type: number
          example: 0.69
        volume:
          type: integer
          example: 12345678
        updatedAt:
          type: string
          format: date-time
...
    StockSearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Stock'
        total:
          type: integer
    StockPriceResponse:
      type: object
      properties:
        symbol:
          type: string
        currentPrice:
          type: number
        previousClose:
          type: number
        changePercent:
          type: number
        updatedAt:
          type: string
          format: date-time

    # ============================================================
    # Technical Indicators Schemas
    # ============================================================
...